+++ back/crawl/index.js
@@ -1,10 +1,66 @@
 const client = require('cheerio-httpcli');
 
-module.exports = new Promise((resolve)=>{
+let crawlresult = {};
+
+new Promise((resolve)=>{
     client.fetch("http://www.skhu.ac.kr/board/boardlist.aspx?bsid=10004&searchBun=51", {},(err, $, res, body)=>{
-    
         const list = $(".left15 #ctl00_ContentPlaceHolder1_ctl00_rptList_ctl00_lblTitle");
         const results = list.find('a').text();
         resolve(results);
     });
+}).then((result)=>{
+    crawlresult.a = result;
 });
+
+new Promise((resolve)=>{
+    client.fetch("http://www.skhu.ac.kr/board/boardlist.aspx?bsid=10005&searchBun=53",{},(err,$,res,body)=>{
+        const list = $(".left15 #ctl00_ContentPlaceHolder1_ctl00_rptList_ctl00_lblTitle");
+        const results = list.find('a').text();
+        resolve(results);
+    });
+}).then((result)=>{
+    crawlresult.b = result;
+});
+
+new Promise((resolve)=>{
+    client.fetch("http://www.skhu.ac.kr/board/boardlist.aspx?bsid=10038&searchBun=89",{},(err,$,res,body)=>{
+        const list = $(".left15 #ctl00_ContentPlaceHolder1_ctl00_rptList_ctl00_lblTitle");
+        const results = list.find('a').text();
+        resolve(results);
+    });
+}).then((result)=>{
+    crawlresult.c = result;
+});
+
+new Promise((resolve)=>{
+    client.fetch("http://www.skhu.ac.kr/board/boardlist.aspx?bsid=10006&searchBun=75",{},(err,$,res,body)=>{
+        const list = $(".left15 #ctl00_ContentPlaceHolder1_ctl00_rptList_ctl00_lblTitle");
+        const results = list.find('a').text();
+        resolve(results);
+    });
+}).then((result)=>{
+    crawlresult.d = result;
+});
+
+new Promise((resolve)=>{
+    client.fetch("http://www.skhu.ac.kr/board/boardlist.aspx?bsid=10007",{},(err,$,res,body)=>{
+        const list = $(".left15 #ctl00_ContentPlaceHolder1_ctl00_rptList_ctl00_lblTitle");
+        const results = list.find('a').text();
+        resolve(results);
+    });
+}).then((result)=>{
+    crawlresult.e = result;
+});
+
+new Promise((resolve)=>{
+    client.fetch("http://www.skhu.ac.kr/board/boardlist.aspx?bsid=10008",{},(err,$,res,body)=>{
+        const list = $(".left15 #ctl00_ContentPlaceHolder1_ctl00_rptList_ctl00_lblTitle");
+        const results = list.find('a').text();
+        resolve(results);
+    });
+}).then((result)=>{
+    crawlresult.f = result;
+});
+
+
+module.exports = crawlresult;
